create formatAllLocationObjects helper
