+++ api/google-maps-services/index.js
@@ -1,10 +1,11 @@
 const {
   googleLocation,
-  googlePlacePhoto
+  googlePlacePhoto,
 } = require("../google-maps-services/client");
 
 module.exports = {
-  googleLocationObject
+  googleLocationObject,
+  formatAllLocationObjects,
 };
 
 async function getPhotoUrl(ref) {
@@ -34,11 +35,30 @@ async function googleLocationObject(location) {
       name: loc.name,
       address: loc.formatted_address,
       phone: loc.formatted_phone_number,
-      icon: photo || loc.icon
+      icon: photo || loc.icon,
     };
 
     return structuredResponse;
   } catch (err) {
     return { message: `Error grabbing location info: ${err.message}` };
   }
+}
+
+async function formatAllLocationObjects(locationsList) {
+  return await Promise.all(
+    locationsList.map(async (l) => {
+      let location = {
+        id: l.id,
+        name: l.name,
+        address: l.address,
+        phone: l.phone,
+        icon: l.icon,
+      };
+      if (!!l.googleId) location = await googleLocationObject(l);
+
+      return location;
+    })
+  ).then((locations) => {
+    return locations;
+  });
 }
